Removed conact form, learned that css leaks between components, fixed card styling on peers and projects, created a functioning and styled contact-modal that now just needs to change to a email submitted modal when sending an email and needs locked by google firebase.
