+++ src/App.js
@@ -4,7 +4,6 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import Footer from './components/Footer';
 import Navigation from './components/Navigation';
 import Resume from './components/Resume/Resume';
-import Contact from './components/Contact/Contact';
 import Home from './components/Home/Home';
 import Peers from './components/Peers/Peers';
 import 'bootstrap'
@@ -19,7 +18,6 @@ function App() {
         <Navigation />
         <Routes>
           <Route path='/' element={<Home />}></Route>
-          <Route path='/Contact' element={<Contact />} />
           <Route path='/Resume' element={<Resume />} />
           <Route path='/Peers' element={<Peers />} />
           <Route path='/Projects' element={<Projects />} />
