--- src/components/Contact/Contact.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { ContactUs } from "./ContactForm";
-
-export default function Contact() {
-    return (
-        <>
-        <h1>Contact Me</h1>
-
-        <ContactUs />
-        </>
-        
-    )
-}
