+++ src/components/Home/Frameworks.jsx
@@ -26,7 +26,7 @@ export default function Frameworks() {
   ]
   const frameworkList = frameworks.map(x => (
     <div class='col-md-5 mb-4'>
-        <img key={x.id} src={x.image} alt={x.framework}></img>
+        <img id='frameworkImg' key={x.id} src={x.image} alt={x.framework}></img>
         <h2>{x.framework}</h2>
         <p>{x.description}</p>
     </div>
