+++ src/components/Home/Languages.jsx
@@ -47,7 +47,7 @@ export default function Languages() {
         
         
             <div class='col-md-5 mb-4'>
-                <img key={x.id} src={x.image} alt={x.language}></img>
+                <img id='languageImg' key={x.id} src={x.image} alt={x.language}></img>
                     <h2>{x.language}</h2>
                     <p>{x.description}</p>
             </div>
